+++ src/demo/dropdown-treeview-select/index.ts
@@ -1,0 +1,1 @@
+export * from './dropdown-treeview-select.module';