+++ src/lib/treeview-config.ts
@@ -2,8 +2,23 @@ import { Injectable } from '@angular/core';
 
 @Injectable()
 export class TreeviewConfig {
-    isShowAllCheckBox = true;
-    isShowFilter = false;
-    isShowCollapseExpand = false;
+    hasAllCheckBox = true;
+    hasFilter = false;
+    hasCollapseExpand = false;
     maxHeight = 500;
+
+    get hasDivider(): boolean {
+        return this.hasFilter || this.hasAllCheckBox || this.hasCollapseExpand;
+    }
+
+    public static create(fields?: {
+        hasAllCheckBox?: boolean,
+        hasFilter?: boolean,
+        hasCollapseExpand?: boolean,
+        maxHeight?: number
+    }): TreeviewConfig {
+        const config = new TreeviewConfig();
+        Object.assign(config, fields);
+        return config;
+    }
 }