+++ src/test.ts
@@ -13,11 +13,11 @@ import {
 } from '@angular/platform-browser-dynamic/testing';
 
 // Unfortunately there's no typing for the `__karma__` variable. Just declare it as any.
-declare var __karma__: any;
-declare var require: any;
+declare const __karma__: any;
+declare const require: any;
 
 // Prevent Karma from running prematurely.
-__karma__.loaded = function () {};
+__karma__.loaded = function () { };
 
 // First, initialize the Angular testing environment.
 getTestBed().initTestEnvironment(
@@ -25,8 +25,8 @@ getTestBed().initTestEnvironment(
   platformBrowserDynamicTesting()
 );
 // Then we find all the tests.
-const context = require.context('./lib', true, /\.spec\.ts$/);
-// const context = require.context('./', true, /dropdown-toggle.directive.spec.ts$/);
+const context = require.context('./', true, /\.spec\.ts$/);
+// const context = require.context('./', true, /treeview-item.component.spec.ts$/);
 // And load the modules.
 context.keys().map(context);
 // Finally, start Karma to run the tests.